backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "thalimage"
version = "0.1.0"
description = "Self-hosted image browser for AI-generated images"
requires-python = ">=3.11"
authors = [
    {name = "Alexandre Avanian", email = "alexandre@151e.ai"},
]
license = {text = "Proprietary"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pillow>=10.0.0",
    "sd-parsers>=0.1.0",
    "piexif>=1.1.3",
    "aiosqlite>=0.20.0",
    "sse-starlette>=2.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "mypy>=1.17.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.12.12",
]

[project.scripts]
thalimage = "thalimage.__main__:main"

[tool.ruff.lint]
extend-select = ["B"]

[tool.ruff.lint.per-file-ignores]
"src/thalimage/api/*.py" = ["B008"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["tests/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
